% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_paper.R
\docType{data}
\name{data_paper}
\alias{data_paper}
\title{Example of the dataset used in the paper.}
\format{
A data frame with 4364 rows and 3 variables:
\describe{
\item{ID}{Id of the people}
\item{Concept}{Concept being described by the person ID}
\item{Property}{A property mentioned for the corresponding concept for the person ID}
}
}
\usage{
data_paper
}
\description{
A real dataset from spanish speakers, translated using automatic algorithms
}
\keyword{datasets}
