% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_test.R
\docType{data}
\name{data_test}
\alias{data_test}
\title{Test example dataset.}
\format{
A data frame with 65 rows and 3 variables:
\describe{
\item{ID}{Id of the people}
\item{Concept}{Concept being described by the person ID}
\item{Property}{A property mentioned for the corresponding concept for the person ID}
}
}
\usage{
data_test
}
\description{
A toy dataset containing the description of ten people, over 3 concept, with multiple properties.
}
\keyword{datasets}
