% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_norms.R
\name{generate_norms}
\alias{generate_norms}
\title{Calculate all the norms from a Conceptual properties}
\usage{
generate_norms(orig_data)
}
\arguments{
\item{orig_data}{a data frame of size nx3 (id, concept, property)}
}
\value{
a data frame with all the estimations
}
\description{
Calculate all the norms from a Conceptual properties
}
\examples{
generate_norms(data_test)
}
