% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property_simulator.R
\name{property_simulator}
\alias{property_simulator}
\title{Simulate properties based on the empricial distribution of the original data and new words with frequency one}
\usage{
property_simulator(orig_data, new_words, number_subjects)
}
\arguments{
\item{orig_data}{a data frame of size nx3 (id, concept, property). The empriical distribution is generated from this data}

\item{new_words}{integer greater than 0, corresponding to the number of words with frequency one that should be added to the empirical distribution}

\item{number_subjects}{number of subjects to be sampled. Each subject with generates new properties}
}
\value{
a vector with the extra number of participant to achieve the especific coverage, and the estimate of the number of unique properties listed by the new amount of suggested people
}
\description{
Simulate properties based on the empricial distribution of the original data and new words with frequency one
}
\examples{
orig_data=data_paper[data_paper[,2]=="Decision",]
property_simulator(orig_data, 84, 15)
}
